faceid/requirements.txt
Flask==3.0.0
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Werkzeug==3.0.1
flask-cors==4.0.0
Pillow==10.1.0
dlib==19.24.6
